+++ Kernel/Modules/AgentTicketBulk.pm
@@ -1,6 +1,6 @@
 # --
 # Copyright (C) 2001-2021 OTRS AG, https://otrs.com/
-# Copyright (C) 2021 Znuny GmbH, https://znuny.org/
+# Copyright (C) 2021-2022 Znuny GmbH, https://znuny.org/
 # --
 # This software comes with ABSOLUTELY NO WARRANTY. For details, see
 # the enclosed file COPYING for license information (GPL). If you
@@ -66,6 +66,29 @@ sub Run {
             # error screen, don't show ticket
             if ( !$Access ) {
                 return $LayoutObject->NoPermission( WithHeader => 'yes' );
+            }
+
+            # set previous owner
+            my @HistoryLines = $TicketObject->HistoryGet(
+                TicketID => $TicketID,
+                UserID   => $Self->{UserID},
+            );
+
+            my @TicketBulkActionHistoryEntries = grep {
+                $_->{HistoryType} eq 'Bulk'
+                    && $_->{Name} eq 'Ticket bulk action.'
+                    && $_->{CreateBy} == $Self->{UserID}
+            } @HistoryLines;
+
+            if (@TicketBulkActionHistoryEntries) {
+                my $NewestTicketBulkActionHistoryEntry = pop @TicketBulkActionHistoryEntries;
+                my $LastOwnerID                        = $NewestTicketBulkActionHistoryEntry->{OwnerID};
+
+                $TicketObject->TicketOwnerSet(
+                    TicketID  => $TicketID,
+                    NewUserID => $LastOwnerID,
+                    UserID    => $Self->{UserID},
+                );
             }
 
             # set unlock
@@ -739,6 +762,13 @@ sub Run {
                         TicketID => $TicketID,
                         Lock     => 'lock',
                         UserID   => $Self->{UserID},
+                    );
+
+                    $TicketObject->HistoryAdd(
+                        Name         => 'Ticket bulk action.',
+                        HistoryType  => 'Bulk',
+                        TicketID     => $TicketID,
+                        CreateUserID => $Self->{UserID},
                     );
 
                     # set user id