+++ scripts/test/Selenium/Agent/AgentTicketBulk.t
@@ -1,6 +1,6 @@
 # --
 # Copyright (C) 2001-2021 OTRS AG, https://otrs.com/
-# Copyright (C) 2021 Znuny GmbH, https://znuny.org/
+# Copyright (C) 2021-2022 Znuny GmbH, https://znuny.org/
 # --
 # This software comes with ABSOLUTELY NO WARRANTY. For details, see
 # the enclosed file COPYING for license information (GPL). If you
@@ -451,6 +451,14 @@ $Selenium->RunTest(
             1,
             "Ticket remind locked after undo in bulk feature - $Tickets[0]->{TicketNumber}"
         );
+
+        # Check for previous owner.
+        $Self->Is(
+            $Tickets[3]->{OwnerID},
+            1,
+            "Ticket owner remain after undo in bulk feature - $Tickets[3]->{TicketNumber}"
+        );
+
         $Selenium->VerifiedRefresh();
 
         # Select test tickets and click on "bulk".